Extract the node config into loomchain/config package
Now the Config type can be reused in other tools like the
validators-tool.
