+++ cmd/loom/config.go
@@ -6,12 +6,8 @@ import (
 	"encoding/json"
 	"io/ioutil"
 	"os"
-	"path"
 	"path/filepath"
-	"strconv"
 	"strings"
-
-	"github.com/loomnetwork/loomchain/store"
 
 	ktypes "github.com/loomnetwork/go-loom/builtin/types/karma"
 	"github.com/pkg/errors"
@@ -23,14 +19,9 @@ import (
 	"github.com/loomnetwork/go-loom/plugin/contractpb"
 	"github.com/loomnetwork/loomchain/builtin/plugins/dpos"
 	"github.com/loomnetwork/loomchain/builtin/plugins/dposv2"
-	"github.com/loomnetwork/loomchain/gateway"
+	"github.com/loomnetwork/loomchain/config"
 	"github.com/loomnetwork/loomchain/plugin"
-	hsmpv "github.com/loomnetwork/loomchain/privval/hsm"
-	receipts "github.com/loomnetwork/loomchain/receipts/handler"
-	registry "github.com/loomnetwork/loomchain/registry/factory"
 	"github.com/loomnetwork/loomchain/vm"
-
-	plasmaConfig "github.com/loomnetwork/loomchain/builtin/plugins/plasma_cash/config"
 )
 
 func decodeHexString(s string) ([]byte, error) {
@@ -41,60 +32,8 @@ func decodeHexString(s string) ([]byte, error) {
 	return hex.DecodeString(s[2:])
 }
 
-type Config struct {
-	RootDir            string
-	DBName             string
-	GenesisFile        string
-	PluginsDir         string
-	QueryServerHost    string
-	EventDispatcherURI string
-	ContractLogLevel   string
-	LogDestination     string
-	LoomLogLevel       string
-	BlockchainLogLevel string
-	Peers              string
-	PersistentPeers    string
-	RPCListenAddress   string
-	ChainID            string
-	RPCProxyPort       int32
-	RPCBindAddress     string
-	// Controls whether or not empty blocks should be generated periodically if there are no txs or
-	// AppHash changes. Defaults to true.
-	CreateEmptyBlocks          bool
-	SessionMaxAccessCount      int64
-	SessionDuration            int64
-	LogStateDB                 bool
-	LogEthDbBatch              bool
-	UseCheckTx                 bool
-	RegistryVersion            int32
-	ReceiptsVersion            int32
-	EVMPersistentTxReceiptsMax uint64
-	TransferGateway            *gateway.TransferGatewayConfig
-	PlasmaCash                 *plasmaConfig.PlasmaCashSerializableConfig
-	// When this setting is enabled Loom EVM accounts are hooked up to the builtin ethcoin Go contract,
-	// which makes it possible to use the payable/transfer features of the EVM to transfer ETH in
-	// Solidity contracts running on the Loom EVM. This setting is disabled by default, which means
-	// all the EVM accounts always have a zero balance.
-	EVMAccountsEnabled bool
-	EVMDebugEnabled    bool
-
-	Oracle        string
-	DeployEnabled bool
-	CallEnabled   bool
-
-	KarmaEnabled         bool
-	KarmaMaxCallCount    int64
-	KarmaSessionDuration int64
-	KarmaMaxDeployCount  int64
-	DPOSVersion          int64
-
-	AppStore *store.AppStoreConfig
-
-	HsmConfig *hsmpv.HsmConfig
-}
-
 // Loads loom.yml from ./ or ./config
-func parseConfig() (*Config, error) {
+func parseConfig() (*config.Config, error) {
 	v := viper.New()
 	v.AutomaticEnv()
 	v.SetEnvPrefix("LOOM")
@@ -105,87 +44,12 @@ func parseConfig() (*Config, error) {
 	v.AddConfigPath("./../../../")
 
 	v.ReadInConfig()
-	conf := DefaultConfig()
+	conf := config.DefaultConfig()
 	err := v.Unmarshal(conf)
 	if err != nil {
 		return nil, err
 	}
 	return conf, err
-}
-
-func (c *Config) fullPath(p string) string {
-	full, err := filepath.Abs(path.Join(c.RootDir, p))
-	if err != nil {
-		panic(err)
-	}
-	return full
-}
-
-func (c *Config) RootPath() string {
-	return c.fullPath(c.RootDir)
-}
-
-func (c *Config) GenesisPath() string {
-	return c.fullPath(c.GenesisFile)
-}
-
-func (c *Config) PluginsPath() string {
-	return c.fullPath(c.PluginsDir)
-}
-
-func DefaultConfig() *Config {
-	cfg := &Config{
-		RootDir:                    ".",
-		DBName:                     "app",
-		GenesisFile:                "genesis.json",
-		PluginsDir:                 "contracts",
-		QueryServerHost:            "tcp://127.0.0.1:9999",
-		RPCListenAddress:           "tcp://0.0.0.0:46657", //TODO this is an ephemeral port in linux, we should move this
-		EventDispatcherURI:         "",
-		ContractLogLevel:           "info",
-		LoomLogLevel:               "info",
-		LogDestination:             "",
-		BlockchainLogLevel:         "error",
-		Peers:                      "",
-		PersistentPeers:            "",
-		ChainID:                    "",
-		RPCProxyPort:               46658,
-		RPCBindAddress:             "tcp://0.0.0.0:46658",
-		CreateEmptyBlocks:          true,
-		LogStateDB:                 false,
-		LogEthDbBatch:              false,
-		UseCheckTx:                 true,
-		RegistryVersion:            int32(registry.RegistryV1),
-		ReceiptsVersion:            int32(receipts.DefaultReceiptStorage),
-		EVMPersistentTxReceiptsMax: receipts.DefaultMaxReceipts,
-		SessionDuration:            600,
-		EVMAccountsEnabled:         false,
-		EVMDebugEnabled:            false,
-
-		Oracle:        "",
-		DeployEnabled: true,
-		CallEnabled:   true,
-
-		KarmaEnabled:         false,
-		KarmaMaxCallCount:    0,
-		KarmaSessionDuration: 0,
-		KarmaMaxDeployCount:  0,
-		DPOSVersion:          1,
-	}
-	cfg.TransferGateway = gateway.DefaultConfig(cfg.RPCProxyPort)
-	cfg.PlasmaCash = plasmaConfig.DefaultConfig()
-	cfg.AppStore = store.DefaultConfig()
-	cfg.HsmConfig = hsmpv.DefaultConfig()
-	return cfg
-}
-
-func (c *Config) QueryServerPort() (int32, error) {
-	hostPort := strings.Split(c.QueryServerHost, ":")
-	port, err := strconv.ParseInt(hostPort[2], 10, 32)
-	if err != nil {
-		return 0, err
-	}
-	return int32(port), nil
 }
 
 type contractConfig struct {
@@ -234,7 +98,7 @@ func marshalInit(pb proto.Message) (json.RawMessage, error) {
 	return json.RawMessage(buf.Bytes()), nil
 }
 
-func defaultGenesis(cfg *Config, validator *loom.Validator) (*genesis, error) {
+func defaultGenesis(cfg *config.Config, validator *loom.Validator) (*genesis, error) {
 	contracts := []contractConfig{
 		{
 			VMTypeName: "plugin",
